requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
black==22.12.0
click==8.1.3
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycrypto==2.6.1
PyJWT==2.6.0
pytz==2022.7.1
redis==4.4.2
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
